+++ FlatFileGenerator.Core/Extensions/DictionaryExtensions.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace FlatFileGenerator.Extensions
+namespace FlatFileGenerator.Core.Extensions
 {
     public static class DictionaryExtensions
     {
@@ -21,6 +21,19 @@ namespace FlatFileGenerator.Extensions
             }
 
             return value;
+        }
+
+        public static TValue GetValueOrDefault<TKey, TValue> (this IDictionary<TKey, TValue> dictionary, TKey key, TValue defaultValue)
+        {
+            TValue value;
+            return dictionary.TryGetValue(key, out value) ? value : defaultValue;
+        }
+
+        public static TValue GetValueOrDefault<TKey, TValue> (this IDictionary<TKey, TValue> dictionary, TKey key, Func<TValue> defaultValueProvider)
+        {
+            TValue value;
+            return dictionary.TryGetValue(key, out value) ? value
+                 : defaultValueProvider();
         }
     }
 }
