+++ FlatFileGenerator.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28307.329
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FlatFileGenerator", "FlatFileGenerator\FlatFileGenerator.csproj", "{86FC503B-6F99-4A2E-AA02-B44DBEA271A3}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FlatFileGenerator", "FlatFileGenerator\FlatFileGenerator.csproj", "{86FC503B-6F99-4A2E-AA02-B44DBEA271A3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FlatFileGenerator.Core", "FlatFileGenerator.Core\FlatFileGenerator.Core.csproj", "{FFC6AA66-D5BE-4706-A323-6E1D2326221A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{86FC503B-6F99-4A2E-AA02-B44DBEA271A3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{86FC503B-6F99-4A2E-AA02-B44DBEA271A3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{86FC503B-6F99-4A2E-AA02-B44DBEA271A3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FFC6AA66-D5BE-4706-A323-6E1D2326221A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FFC6AA66-D5BE-4706-A323-6E1D2326221A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FFC6AA66-D5BE-4706-A323-6E1D2326221A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FFC6AA66-D5BE-4706-A323-6E1D2326221A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
